[129/webpack] organize into more files
