+++ BookReader/plugins/plugin.tts.js
@@ -1,1 +1,1 @@
-!function(t){var e={};function s(n){if(e[n])return e[n].exports;var r=e[n]={i:n,l:!1,exports:{}};return t[n].call(r.exports,r,r.exports,s),r.l=!0,r.exports}s.m=t,s.c=e,s.d=function(t,e,n){s.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},s.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},s.t=function(t,e){if(1&e&&(t=s(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(s.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)s.d(n,r,function(e){return t[e]}.bind(null,r));return n},s.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return s.d(e,"a",e),e},s.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},s.p="",s(s.s=0)}([function(t,e,s){"use strict";s.r(e);class n{pull(){return null}map(t){return new o(this,t)}filter(t){return new i(this,t)}buffer(t){return new u(this,t)}flatten(){return new a(this)}static range(t,e){return new r(t,e)}}class r extends n{constructor(t,e){super(),this.start=t,this.end=e,this.cur=t}pull(){return this.cur<=this.end?Promise.resolve({value:this.cur++,done:!1}):Promise.resolve({value:null,done:!0})}}class o extends n{constructor(t,e){super(),this.parent=t,this.mapper=e}pull(){return this.parent.pull().then(t=>{if(t.done)return Promise.resolve({value:null,done:!0});var e=this.mapper(t.value);return e instanceof Promise||(e=Promise.resolve(e)),e.then(t=>({value:t,done:!1}))})}}class i extends n{constructor(t,e){super(),this.parent=t,this.predicate=e}pull(){return this.parent.pull().then(t=>t.done?Promise.resolve({value:null,done:!0}):this.predicate(t.value)?Promise.resolve(t):this.pull())}}class a extends n{constructor(t){super(),this.parent=t,this._lastResult=[],this._inProgress=null}pull(){return this._lastResult.length?Promise.resolve({value:this._lastResult.shift(),done:!1}):this._inProgress?this._inProgress.then(()=>this.pull()):(this._inProgress=this.parent.pull().then(t=>t.done?Promise.resolve({value:null,done:!0}):(this._lastResult=t.value,this._inProgress=null,this.pull())),this._inProgress)}}class u extends n{constructor(t,e){super(),this.parent=t,this._bufferSize=e,this._promiseBuffer=[]}pull(){for(var t=Math.max(0,this._bufferSize-this._promiseBuffer.length),e=0;e<t;e++)this._promiseBuffer.push(this.parent.pull());return this._promiseBuffer.shift()}}jQuery.extend(BookReader.defaultOptions,{server:"ia600609.us.archive.org",bookPath:"",enableTtsPlugin:!0});class l{constructor(t){this.playing=!1,this.opts=t,this.chunkStream=null,this.isSupported=!1}init(){return null}start(t,e){this.playing=!0,this.opts.onLoadingStart(),this.chunkStream=n.range(t,e-1).map(this.fetchPageChunks.bind(this)).buffer(2).flatten(),this.step()}stop(){this.playing=!1,this.chunkStream=null}step(){this.getPlayStream().pull().then(t=>(this.opts.onLoadingComplete(),this.opts.beforeChunkPlay(t.value),this.playChunk(t.value))).then(()=>{if(this.playing)return this.step()})}getPlayStream(){return null}playChunk(t){return null}fetchPageChunks(t){return $.ajax({type:"GET",url:"https://"+this.opts.server+"/BookReader/BookReaderGetTextWrapper.php",dataType:"jsonp",data:{path:this.opts.bookPath+"_djvu.xml",page:t}}).then((function(e){return e.map(e=>({leafIndex:t,text:e[0],lineRects:e.slice(1)}))}))}}class h extends l{constructor(t){super(t),this.audioFormat=$.browser.mozilla?"ogg":"mp3",this.isSupported="undefined"!=typeof soundManager&&soundManager.supported()}init(){this.isSupported&&soundManager.setup({debugMode:!1,preferFlash:!0,url:"/bookreader/BookReader/soundmanager/swf",useHTML5Audio:!0,flashVersion:9})}start(t,e){navigator.userAgent.match(/mobile/i)&&soundManager.createSound({url:this.getSoundUrl(" ")}).play(),super.start(t,e)}stop(){this.playStream=null,soundManager.stopAll(),super.stop()}getPlayStream(){return this.playStream=this.playStream||this.chunkStream.map(this.fetchChunkSound.bind(this)).buffer(2),this.playStream}playChunk(t){return new Promise(e=>t.sound.play({onfinish:e})).then(()=>t.sound.destruct())}fetchChunkSound(t){return new Promise(e=>{function s(s){e($.extend(t,{sound:s}))}soundManager.createSound({url:this.getSoundUrl(t.text),onload:function(){s(this)},onready:function(){s(this)},onbufferchange:function(){this.isBuffering||s(this)}}).load()})}getSoundUrl(t){return"https://"+this.opts.server+"/BookReader/BookReaderGetTTS.php?string="+encodeURIComponent(t)+"&format=."+this.audioFormat}}var p;BookReader.prototype.setup=(p=BookReader.prototype.setup,function(t){p.call(this,t),this.options.enableTtsPlugin&&(this.ttsEngine=new h({server:t.server,bookPath:t.bookPath,onLoadingStart:this.showProgressPopup.bind(this,"Loading audio..."),onLoadingComplete:this.removeProgressPopup.bind(this),beforeChunkPlay:this.ttsBeforeChunkPlay.bind(this)}),this.ttsHilites=[])}),BookReader.prototype.init=function(t){return function(){this.options.enableTtsPlugin&&this.bind(BookReader.eventNames.PostInit,(function(t,e){jIcons=e.$(".BRicon.read").click((function(t){return e.ttsToggle(),!1})),e.ttsEngine.init()})),t.call(this)}}(BookReader.prototype.init),BookReader.prototype.buildMobileDrawerElement=function(t){return function(){var e=t.call(this);return this.options.enableTtsPlugin&&this.ttsEngine.isSupported&&e.find(".BRmobileMenu__moreInfoRow").after($('    <li>      <span>        <span class="DrawerIconWrapper "><img class="DrawerIcon" src="'+this.imagesBaseURL+"icon_speaker_open.svg\" alt=\"info-speaker\"/></span>        Read Aloud      </span>      <div>        <span class='larger'>Press to toggle read aloud</span> <br/>        <button class='BRicon read'></button>      </div>    </li>")),e}}(BookReader.prototype.buildMobileDrawerElement),BookReader.prototype.initNavbar=function(t){return function(){var e=t.call(this);return this.options.enableTtsPlugin&&this.ttsEngine.isSupported&&$("<button class='BRicon read js-tooltip'></button>").insertAfter(e.find(".BRpage .BRicon.thumb")),e}}(BookReader.prototype.initNavbar),BookReader.prototype.ttsToggle=function(){this.autoStop&&this.autoStop(),this.ttsEngine.playing?this.ttsStop():this.ttsStart()},BookReader.prototype.ttsStart=function(){this.constModeThumb==this.mode&&this.switchMode(this.constMode1up),this.$(".BRicon.read").addClass("unread"),this.ttsEngine.start(this.currentIndex(),this.getNumLeafs())},BookReader.prototype.ttsStop=function(){this.ttsEngine.playing&&(this.$(".BRicon.read").removeClass("unread"),this.ttsEngine.stop(),this.ttsRemoveHilites(),this.removeProgressPopup())},BookReader.prototype.ttsBeforeChunkPlay=function(t){this.ttsMaybeFlip(t.leafIndex),this.ttsHighlightChunk(t),this.ttsScrollToChunk(t)},BookReader.prototype.ttsMaybeFlip=function(t){var e=this.constMode2up==this.mode,s=t>this.twoPage.currentIndexR,n=$.Deferred();return e&&s?(this.animationFinishedCallback=n.resolve.bind(n),this.next()):n.resolve(),n.promise()},BookReader.prototype.ttsHighlightChunk=function(t){this.ttsRemoveHilites(),this.constMode2up==this.mode?this.ttsHilite2UP(t):this.ttsHilite1UP(t)},BookReader.prototype.ttsScrollToChunk=function(t){if(this.constMode1up==this.mode){var e,s=0;for(e=0;e<t.leafIndex;e++)s+=parseInt(this._getPageHeight(e)/this.reduce)+this.padding;var n=t.lineRects[0][3],r=t.lineRects[t.lineRects.length-1][1],o=s+n/this.reduce,i=s+r/this.reduce;soundManager.debugMode&&console.log("leafTop = "+s+" topOfFirstChunk = "+o+" botOfLastChunk = "+i);var a=this.refs.$brContainer.prop("scrollTop"),u=a+this.refs.$brContainer.height();soundManager.debugMode&&console.log("containerTop = "+a+" containerBot = "+u),(o<a||i>u)&&this.refs.$brContainer.stop(!0).animate({scrollTop:o},"fast")}},BookReader.prototype.ttsHilite1UP=function(t){for(var e=0;e<t.lineRects.length;e++){var s=t.lineRects[e][0],n=t.lineRects[e][1],r=t.lineRects[e][2],o=t.lineRects[e][3],i=document.createElement("div");this.ttsHilites.push(i),$(i).prop("className","BookReaderSearchHilite").appendTo(this.$(".pagediv"+t.leafIndex)),$(i).css({width:(r-s)/this.reduce+"px",height:(n-o)/this.reduce+"px",left:s/this.reduce+"px",top:o/this.reduce+"px"})}},BookReader.prototype.ttsHilite2UP=function(t){for(var e=0;e<t.lineRects.length;e++){var s=t.lineRects[e][0],n=t.lineRects[e][1],r=t.lineRects[e][2],o=t.lineRects[e][3],i=document.createElement("div");this.ttsHilites.push(i),$(i).prop("className","BookReaderSearchHilite").css("zIndex",3).appendTo(this.refs.$brTwoPageView),this.setHilightCss2UP(i,t.leafIndex,s,r,o,n)}},BookReader.prototype.ttsRemoveHilites=function(){$(this.ttsHilites).remove(),this.ttsHilites=[]}}]);
+!function(t){var e={};function s(n){if(e[n])return e[n].exports;var r=e[n]={i:n,l:!1,exports:{}};return t[n].call(r.exports,r,r.exports,s),r.l=!0,r.exports}s.m=t,s.c=e,s.d=function(t,e,n){s.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},s.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},s.t=function(t,e){if(1&e&&(t=s(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(s.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)s.d(n,r,function(e){return t[e]}.bind(null,r));return n},s.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return s.d(e,"a",e),e},s.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},s.p="",s(s.s=0)}([function(t,e,s){"use strict";s.r(e);class n{pull(){return null}map(t){return new o(this,t)}filter(t){return new i(this,t)}buffer(t){return new u(this,t)}flatten(){return new a(this)}static range(t,e){return new r(t,e)}}class r extends n{constructor(t,e){super(),this.start=t,this.end=e,this.cur=t}pull(){return this.cur<=this.end?Promise.resolve({value:this.cur++,done:!1}):Promise.resolve({value:null,done:!0})}}class o extends n{constructor(t,e){super(),this.parent=t,this.mapper=e}pull(){return this.parent.pull().then(t=>{if(t.done)return Promise.resolve({value:null,done:!0});var e=this.mapper(t.value);return e instanceof Promise||(e=Promise.resolve(e)),e.then(t=>({value:t,done:!1}))})}}class i extends n{constructor(t,e){super(),this.parent=t,this.predicate=e}pull(){return this.parent.pull().then(t=>t.done?Promise.resolve({value:null,done:!0}):this.predicate(t.value)?Promise.resolve(t):this.pull())}}class a extends n{constructor(t){super(),this.parent=t,this._lastResult=[],this._inProgress=null}pull(){return this._lastResult.length?Promise.resolve({value:this._lastResult.shift(),done:!1}):this._inProgress?this._inProgress.then(()=>this.pull()):(this._inProgress=this.parent.pull().then(t=>t.done?Promise.resolve({value:null,done:!0}):(this._lastResult=t.value,this._inProgress=null,this.pull())),this._inProgress)}}class u extends n{constructor(t,e){super(),this.parent=t,this._bufferSize=e,this._promiseBuffer=[]}pull(){for(var t=Math.max(0,this._bufferSize-this._promiseBuffer.length),e=0;e<t;e++)this._promiseBuffer.push(this.parent.pull());return this._promiseBuffer.shift()}}class l{constructor(t){this.playing=!1,this.opts=t,this.chunkStream=null,this.isSupported=!1}init(){return null}start(t,e){this.playing=!0,this.opts.onLoadingStart(),this.chunkStream=n.range(t,e-1).map(this.fetchPageChunks.bind(this)).buffer(2).flatten(),this.step()}stop(){this.playing=!1,this.chunkStream=null}step(){this.getPlayStream().pull().then(t=>(this.opts.onLoadingComplete(),this.opts.beforeChunkPlay(t.value),this.playChunk(t.value))).then(()=>{if(this.playing)return this.step()})}getPlayStream(){return null}playChunk(t){return null}fetchPageChunks(t){return $.ajax({type:"GET",url:"https://"+this.opts.server+"/BookReader/BookReaderGetTextWrapper.php",dataType:"jsonp",data:{path:this.opts.bookPath+"_djvu.xml",page:t}}).then((function(e){return e.map(e=>({leafIndex:t,text:e[0],lineRects:e.slice(1)}))}))}}class h extends l{constructor(t){super(t),this.audioFormat=$.browser.mozilla?"ogg":"mp3",this.isSupported="undefined"!=typeof soundManager&&soundManager.supported()}init(){this.isSupported&&soundManager.setup({debugMode:!1,preferFlash:!0,url:"/bookreader/BookReader/soundmanager/swf",useHTML5Audio:!0,flashVersion:9})}start(t,e){navigator.userAgent.match(/mobile/i)&&soundManager.createSound({url:this.getSoundUrl(" ")}).play(),super.start(t,e)}stop(){this.playStream=null,soundManager.stopAll(),super.stop()}getPlayStream(){return this.playStream=this.playStream||this.chunkStream.map(this.fetchChunkSound.bind(this)).buffer(2),this.playStream}playChunk(t){return new Promise(e=>t.sound.play({onfinish:e})).then(()=>t.sound.destruct())}fetchChunkSound(t){return new Promise(e=>{function s(s){e($.extend(t,{sound:s}))}soundManager.createSound({url:this.getSoundUrl(t.text),onload:function(){s(this)},onready:function(){s(this)},onbufferchange:function(){this.isBuffering||s(this)}}).load()})}getSoundUrl(t){return"https://"+this.opts.server+"/BookReader/BookReaderGetTTS.php?string="+encodeURIComponent(t)+"&format=."+this.audioFormat}}var p;jQuery.extend(BookReader.defaultOptions,{server:"ia600609.us.archive.org",bookPath:"",enableTtsPlugin:!0}),BookReader.prototype.setup=(p=BookReader.prototype.setup,function(t){p.call(this,t),this.options.enableTtsPlugin&&(this.ttsEngine=new h({server:t.server,bookPath:t.bookPath,onLoadingStart:this.showProgressPopup.bind(this,"Loading audio..."),onLoadingComplete:this.removeProgressPopup.bind(this),beforeChunkPlay:this.ttsBeforeChunkPlay.bind(this)}),this.ttsHilites=[])}),BookReader.prototype.init=function(t){return function(){this.options.enableTtsPlugin&&this.bind(BookReader.eventNames.PostInit,(function(t,e){jIcons=e.$(".BRicon.read").click((function(t){return e.ttsToggle(),!1})),e.ttsEngine.init()})),t.call(this)}}(BookReader.prototype.init),BookReader.prototype.buildMobileDrawerElement=function(t){return function(){var e=t.call(this);return this.options.enableTtsPlugin&&this.ttsEngine.isSupported&&e.find(".BRmobileMenu__moreInfoRow").after($('    <li>      <span>        <span class="DrawerIconWrapper "><img class="DrawerIcon" src="'+this.imagesBaseURL+"icon_speaker_open.svg\" alt=\"info-speaker\"/></span>        Read Aloud      </span>      <div>        <span class='larger'>Press to toggle read aloud</span> <br/>        <button class='BRicon read'></button>      </div>    </li>")),e}}(BookReader.prototype.buildMobileDrawerElement),BookReader.prototype.initNavbar=function(t){return function(){var e=t.call(this);return this.options.enableTtsPlugin&&this.ttsEngine.isSupported&&$("<button class='BRicon read js-tooltip'></button>").insertAfter(e.find(".BRpage .BRicon.thumb")),e}}(BookReader.prototype.initNavbar),BookReader.prototype.ttsToggle=function(){this.autoStop&&this.autoStop(),this.ttsEngine.playing?this.ttsStop():this.ttsStart()},BookReader.prototype.ttsStart=function(){this.constModeThumb==this.mode&&this.switchMode(this.constMode1up),this.$(".BRicon.read").addClass("unread"),this.ttsEngine.start(this.currentIndex(),this.getNumLeafs())},BookReader.prototype.ttsStop=function(){this.ttsEngine.playing&&(this.$(".BRicon.read").removeClass("unread"),this.ttsEngine.stop(),this.ttsRemoveHilites(),this.removeProgressPopup())},BookReader.prototype.ttsBeforeChunkPlay=function(t){this.ttsMaybeFlip(t.leafIndex),this.ttsHighlightChunk(t),this.ttsScrollToChunk(t)},BookReader.prototype.ttsMaybeFlip=function(t){var e=this.constMode2up==this.mode,s=t>this.twoPage.currentIndexR,n=$.Deferred();return e&&s?(this.animationFinishedCallback=n.resolve.bind(n),this.next()):n.resolve(),n.promise()},BookReader.prototype.ttsHighlightChunk=function(t){this.ttsRemoveHilites(),this.constMode2up==this.mode?this.ttsHilite2UP(t):this.ttsHilite1UP(t)},BookReader.prototype.ttsScrollToChunk=function(t){if(this.constMode1up==this.mode){var e,s=0;for(e=0;e<t.leafIndex;e++)s+=parseInt(this._getPageHeight(e)/this.reduce)+this.padding;var n=t.lineRects[0][3],r=t.lineRects[t.lineRects.length-1][1],o=s+n/this.reduce,i=s+r/this.reduce;soundManager.debugMode&&console.log("leafTop = "+s+" topOfFirstChunk = "+o+" botOfLastChunk = "+i);var a=this.refs.$brContainer.prop("scrollTop"),u=a+this.refs.$brContainer.height();soundManager.debugMode&&console.log("containerTop = "+a+" containerBot = "+u),(o<a||i>u)&&this.refs.$brContainer.stop(!0).animate({scrollTop:o},"fast")}},BookReader.prototype.ttsHilite1UP=function(t){for(var e=0;e<t.lineRects.length;e++){var s=t.lineRects[e][0],n=t.lineRects[e][1],r=t.lineRects[e][2],o=t.lineRects[e][3],i=document.createElement("div");this.ttsHilites.push(i),$(i).prop("className","BookReaderSearchHilite").appendTo(this.$(".pagediv"+t.leafIndex)),$(i).css({width:(r-s)/this.reduce+"px",height:(n-o)/this.reduce+"px",left:s/this.reduce+"px",top:o/this.reduce+"px"})}},BookReader.prototype.ttsHilite2UP=function(t){for(var e=0;e<t.lineRects.length;e++){var s=t.lineRects[e][0],n=t.lineRects[e][1],r=t.lineRects[e][2],o=t.lineRects[e][3],i=document.createElement("div");this.ttsHilites.push(i),$(i).prop("className","BookReaderSearchHilite").css("zIndex",3).appendTo(this.refs.$brTwoPageView),this.setHilightCss2UP(i,t.leafIndex,s,r,o,n)}},BookReader.prototype.ttsRemoveHilites=function(){$(this.ttsHilites).remove(),this.ttsHilites=[]}}]);
