+++ src/js/plugins/tts/plugin.tts.js
@@ -2,7 +2,7 @@
  * Plugin for Text to Speech in BookReader
  */
 
-import AsyncStream from '../utils/AsyncStream.js';
+import FestivalTTSEngine from './FestivalTTSEngine.js';
 
 // Default options for TTS
 jQuery.extend(BookReader.defaultOptions, {
@@ -10,249 +10,6 @@ jQuery.extend(BookReader.defaultOptions, {
     bookPath: '',
     enableTtsPlugin: true,
 });
-
-/**
- * @typedef {[number, number, number, number]} DJVURect
- * coords are in l,b,r,t order
- */
-
-/**
- * @typedef {Object} PageChunk
- * @property {Number} leafIndex
- * @property {String} text
- * @property {DJVURect[]} lineRects
- */
-
-/**
- * @typedef {Object} TTSEngineOptions 
- * @property {String} server
- * @property {String} bookPath
- * @property {Function} onLoadingStart
- * @property {Function} onLoadingComplete
- * @property {function(PageChunk): void} beforeChunkPlay
- */
-
-/**
- * @template TSound extra properties added to PageChunk to allow playing
- */
-class AbstractTTSEngine {
-    /**
-     * @protected
-     * @param {TTSEngineOptions} options 
-     */
-    constructor(options) {
-        this.playing = false;
-        this.opts = options;
-        /** @type {AsyncStream<PageChunk>} */
-        this.chunkStream = null;
-        /** Different engines should overwrite this */
-        this.isSupported = false;
-    }
-
-    /** @abstract */
-    init() { return null; }
-
-    /**
-     * @param {number} leafIndex
-     * @param {number} numLeafs total number of leafs in the current book
-     */
-    start(leafIndex, numLeafs) {
-        this.playing = true;
-        this.opts.onLoadingStart();
-
-        /** @type {AsyncStream<PageChunk>} */
-        this.chunkStream = AsyncStream.range(leafIndex, numLeafs-1)
-        .map(this.fetchPageChunks.bind(this))
-        .buffer(2)
-        .flatten();
-
-        this.step();
-    }
-
-    stop() {
-        this.playing = false;
-        this.chunkStream = null;
-    }
-
-    /**
-     * @private
-     */
-    step() {
-        this.getPlayStream().pull()
-        .then(item => {
-            this.opts.onLoadingComplete();
-            this.opts.beforeChunkPlay(item.value);
-            return this.playChunk(item.value);
-        })
-        .then(() => {
-            if (this.playing) return this.step();
-        });
-    }
-
-    /**
-
-     * @abstract
-     * @return {AbstractStream<PageChunk & TSound>}
-     */
-    getPlayStream() { return null; }
-
-    /**
-     * @abstract
-     * @param {PageChunk & TSound} chunk
-     * @return {PromiseLike} promise called once playing finished
-     */
-    playChunk(chunk) { return null; }
-
-    /**
-     * @private
-     * Gets the text on a page with given index
-     * @param {Number} leafIndex
-     * @return {PromiseLike<Array<PageChunk>>}
-     */
-    fetchPageChunks(leafIndex) {
-        return $.ajax({
-            type: 'GET',
-            url: 'https://'+this.opts.server+'/BookReader/BookReaderGetTextWrapper.php',
-            dataType:'jsonp',
-            data: {
-                path: this.opts.bookPath+'_djvu.xml',
-                page: leafIndex
-            }
-        })
-        .then(
-            /** @param {Array<[String, ...Array<DJVURect>]>} chunks */
-            function (chunks) {
-                return chunks.map(c => {
-                    return {
-                        leafIndex,
-                        text: c[0],
-                        lineRects: c.slice(1)
-                    };
-                });
-            }
-        );
-    }
-}
-
-/**
- * @extends AbstractTTSEngine<{ sound: SMSound }>
- * TTS using Festival endpoint
- **/
-class FestivalTTSEngine extends AbstractTTSEngine {
-    /**
-     * @param {TTSEngineOptions} options 
-     */
-    constructor(options) {
-        super(options);
-        /** @type {'mp3' | 'ogg'} format of audio to get */
-        this.audioFormat = $.browser.mozilla ? 'ogg' : 'mp3';
-        /** @type {Boolean} Whether this tts engine can run */
-        this.isSupported = typeof(soundManager) !== 'undefined' && soundManager.supported();
-    }
-
-    /** @override */
-    init() {
-        if (this.isSupported) {
-            // setup sound manager
-            soundManager.setup({
-                debugMode: false,
-                // Note, there's a bug in Chrome regarding range requests.
-                // Flash is used as a workaround.
-                // See https://bugs.chromium.org/p/chromium/issues/detail?id=505707
-                preferFlash: true,
-                url: '/bookreader/BookReader/soundmanager/swf',
-                useHTML5Audio: true,
-                //flash 8 version of swf is buggy when calling play() on a sound that is still loading
-                flashVersion: 9
-            });
-        }
-    }
-
-    /**
-     * @override
-     * @param {number} leafIndex
-     * @param {number} numLeafs total number of leafs in the current book
-     */
-    start(leafIndex, numLeafs) {
-        if (navigator.userAgent.match(/mobile/i)) {
-            // HACK for iOS. Security restrictions require playback to be triggered
-            // by a user click/touch. This intention gets lost in the ajax callback
-            // above, but for some reason, if we start the audio here, it works
-            soundManager.createSound({url: this.getSoundUrl(' ')}).play();
-        }
-        
-        super.start(leafIndex, numLeafs);
-    }
-
-    /** @override */
-    stop() {
-        this.playStream = null;
-        soundManager.stopAll();
-        super.stop();
-    }
-
-    /** @override */
-    getPlayStream() {
-        this.playStream = this.playStream || this.chunkStream
-        .map(this.fetchChunkSound.bind(this))
-        .buffer(2);
-
-        return this.playStream;
-    }
-
-    /**
-     * @override
-     * @param {PageChunk & { sound: SMSound }} chunk
-     * @return {PromiseLike}
-     */
-    playChunk(chunk) {
-        return new Promise(res => chunk.sound.play({ onfinish: res }))
-        .then(() => chunk.sound.destruct());
-    }
-
-    /**
-     * @private
-     * @param {PageChunk} pageChunk
-     * @return {PromiseLike<PageChunk & { sound: SMSound }>}
-     */
-    fetchChunkSound(pageChunk) {
-        return new Promise(res => {
-            function resolve(sound) {
-                res($.extend(pageChunk, { sound }))
-            }
-            soundManager.createSound({
-                url: this.getSoundUrl(pageChunk.text),
-
-                // Many soundManger2 callbacks are broken when using HTML5 audio.
-                // whileloading: broken on safari, worked in FF4, but broken on FireFox 5
-                // onload: fires on safari, but *after* the sound starts playing, and does not fire in FF or IE9
-                // onbufferchange: fires in FF5 using HTML5 audio, but not in safari using flash audio
-                // whileplaying: fires everywhere
-                onload: function() { resolve(this); },
-                //fires in FF and IE9
-                onready: function() { resolve(this); },
-                //fires in safari...
-                onbufferchange: function() {
-                    if (!this.isBuffering) {
-                        resolve(this);
-                    }
-                }
-            }).load();
-        });
-    }
-
-    /**
-     * @private
-     * Get URL for audio that says this text
-     * @param {String} dataString the thing to say
-     * @return {String} url
-     */
-    getSoundUrl(dataString) {
-        return 'https://'+this.opts.server+'/BookReader/BookReaderGetTTS.php?string='
-                  + encodeURIComponent(dataString)
-                  + '&format=.'+this.audioFormat;
-    }
-}
 
 // Extend the constructor to add TTS properties
 BookReader.prototype.setup = (function (super_) {
